fix: ability to take into account NaNs in input data
This commit also changes the DataCalibrationLikelihood so that the input antenna S11 model is a callable rather than array, which makes it more flexible for having different frequency arrays for the antenna versus calibration.

BREAKING CHANGE: Input S11 model for the antenna should now be a callable function returning a complex array given an array of frequency.